move create_truncated_row_items to libcommon
