+++ services/worker/src/worker/job_runners/split/first_rows_from_parquet.py
@@ -18,12 +18,11 @@ from libcommon.storage import StrPath
 from libcommon.storage_client import StorageClient
 from libcommon.utils import get_json_size
 from libcommon.viewer_utils.features import to_features_list
-from libcommon.viewer_utils.rows import transform_rows
+from libcommon.viewer_utils.rows import create_truncated_row_items, transform_rows
 
 from worker.config import AppConfig, FirstRowsConfig
 from worker.dtos import CompleteJobResult, SplitFirstRowsResponse
 from worker.job_runners.split.split_job_runner import SplitJobRunner
-from worker.utils import create_truncated_row_items
 
 
 def compute_first_rows_response(
