+++ services/worker/src/worker/job_runners/split/first_rows_from_streaming.py
@@ -24,16 +24,12 @@ from libcommon.exceptions import (
 from libcommon.storage_client import StorageClient
 from libcommon.utils import get_json_size
 from libcommon.viewer_utils.features import to_features_list
-from libcommon.viewer_utils.rows import transform_rows
+from libcommon.viewer_utils.rows import create_truncated_row_items, transform_rows
 
 from worker.config import AppConfig, FirstRowsConfig
 from worker.dtos import CompleteJobResult, SplitFirstRowsResponse
 from worker.job_runners.split.split_job_runner import SplitJobRunnerWithDatasetsCache
-from worker.utils import (
-    create_truncated_row_items,
-    get_rows_or_raise,
-    resolve_trust_remote_code,
-)
+from worker.utils import get_rows_or_raise, resolve_trust_remote_code
 
 
 def compute_first_rows_response(
